aaa: Using extensions API
This patch  uses the extensions API
(Internal/KerberosLdap) in a gradual manner
1. Changing LoginBaseCommand and AuthenticationFilter to work directly with the new API
2. Introducing AuthzUtils as helper class for authorization
3. Changing KerberosLdapDirectory to implement Extension, and base all the code
on search of principals and groups
4. Changing AuthenticationProfileRepository to work with Extension proxies
5. Changing extension manager

Topic: AAA

+++ backend/manager/modules/common/src/main/java/org/ovirt/engine/core/common/config/ConfigValues.java
@@ -728,7 +728,7 @@ public enum ConfigValues {
     @DefaultValueAttribute("1000")
     LdapQueryPageSize,
 
-    @TypeConverterAttribute(Long.class)
+    @TypeConverterAttribute(Integer.class)
     @DefaultValueAttribute("100")
     MaxLDAPQueryPartsNumber,
